split command core to main/command/
