+++ main/CommandStrings.scala
@@ -3,37 +3,15 @@
  */
 package sbt
 
-import complete.HistoryCommands
-import scala.annotation.tailrec
-
-import java.io.File
-import Path._
-
-object CommandSupport
+object CommandStrings
 {
-	def logger(s: State) = globalLogging(s).full
-	def globalLogging(s: State) = s get Keys.globalLogging getOrElse error("Global logging misconfigured")
+	@deprecated("Use the `log` member of a State instance directly.", "0.12.0")
+	def logger(s: State) = s.log
 
-	// slightly better fallback in case of older launcher
-	def bootDirectory(state: State): File =
-		try { state.configuration.provider.scalaProvider.launcher.bootDirectory }
-		catch { case e: NoSuchMethodError => new File(".").getAbsoluteFile }
-
-	private def canRead = (_: File).canRead
-	def notReadable(files: Seq[File]): Seq[File] = files filterNot canRead
-	def readable(files: Seq[File]): Seq[File] = files filter canRead
-	def sbtRCs(s: State): Seq[File] =
-		(Path.userHome / sbtrc) ::
-		(s.baseDir / sbtrc asFile) ::
-		Nil
-
-	def readLines(files: Seq[File]): Seq[String] = files flatMap (line => IO.readLines(line)) flatMap processLine
-	def processLine(s: String) = { val trimmed = s.trim; if(ignoreLine(trimmed)) None else Some(trimmed) }
-	def ignoreLine(s: String) = s.isEmpty || s.startsWith("#")
+	@deprecated("Use the `globalLogging` member of a State instance directly.", "0.12.0")
+	def globalLogging(s: State) = s.globalLogging
 
 	/** The prefix used to identify a request to execute the remaining input on source changes.*/
-	val ContinuousExecutePrefix = "~"
-	val HelpCommand = "help"
 	val AboutCommand = "about"
 	val TasksCommand = "tasks"
 	val ProjectCommand = "project"
@@ -41,8 +19,14 @@ object CommandSupport
 	val ShowCommand = "show"
 	val BootCommand = "boot"
 
-	val Exit = "exit"
-	val Quit = "quit"
+	@deprecated("Moved to BasicCommandStrings", "0.12.0")
+	val ContinuousExecutePrefix = BasicCommandStrings.ContinuousExecutePrefix
+
+	@deprecated("Moved to BasicCommandStrings", "0.12.0")
+	val Exit = BasicCommandStrings.Exit
+
+	@deprecated("Moved to BasicCommandStrings", "0.12.0")
+	val Quit = BasicCommandStrings.Quit
 
 	val EvalCommand = "eval"
 	val evalBrief = (EvalCommand + " <expression>", "Evaluates the given Scala expression and prints the result and type.")
@@ -128,9 +112,8 @@ SetCommand + """ <setting-expression>
 	def sessionBrief = (SessionCommand + " <session-command>", "Manipulates session settings.  For details, run 'help " + SessionCommand + "'.")
 
 	/** The command name to terminate the program.*/
-	val TerminateAction: String = Exit
-
-	def continuousBriefHelp = (ContinuousExecutePrefix + " <command>", "Executes the specified command whenever source files change.")
+	@deprecated("Moved to BasicCommandStrings", "0.12.0")
+	val TerminateAction: String = BasicCommandStrings.TerminateAction
 
 	def tasksPreamble = """
 This is a list of tasks defined for the current project.
@@ -139,11 +122,6 @@ Tasks produce values.  Use the 'show' command to run the task and print the resu
 
 	def tasksBrief = "Displays the tasks defined for the current project."
 	def tasksDetailed = "Displays the tasks defined directly or indirectly for the current project."
-
-	def helpBrief = (HelpCommand + " [command]*", "Displays this help message or prints detailed help on requested commands.")
-	def helpDetailed = """
-If an argument is provided, this prints detailed help for that command.
-Otherwise, this prints a help summary."""
 
 	def aboutBrief = "Displays basic information about sbt and the build."
 	def aboutDetailed = aboutBrief
@@ -175,126 +153,23 @@ ProjectCommand +
 	def projectsBrief = projectsDetailed
 	def projectsDetailed = "Displays the names of available projects."
 
-	def historyHelp = Help.briefDetail(HistoryCommands.descriptions)
-
-	def exitBrief = "Terminates the build."
-
 	def sbtrc = ".sbtrc"
 
-	def ReadCommand = "<"
-	def ReadFiles = " file1 file2 ..."
-	def ReadBrief = (ReadCommand + " <file>*", "Reads command lines from the provided files.")
-	def ReadDetailed =
-ReadCommand + ReadFiles + """
-
-	Reads the lines from the given files and inserts them as commands.
-	All empty lines and lines that start with '#' are ignored.
-	If a file does not exist or is not readable, this command fails.
-
-	All the lines from all the files are read before any of the commands
-	  are executed. Thus, if any file is not readable, none of commands
-	  from any of the files (even the existing ones) will be run.
-
-	You probably need to escape this command if entering it at your shell."""
-
-	def ApplyCommand = "apply"
-	def ApplyBrief = (ApplyCommand + " <module-name>*", ApplyDetailed)
-	def ApplyDetailed = "Transforms the current State by calling <module-name>.apply(currentState) for each listed."
+	@deprecated("Moved to BasicCommandStrings", "0.12.0")
+	def ReadCommand = BasicCommandStrings.ReadCommand
 
 	def DefaultsCommand = "add-default-commands"
 	def DefaultsBrief = (DefaultsCommand, DefaultsDetailed)
 	def DefaultsDetailed = "Registers default built-in commands"
 
+	@deprecated("Moved to BasicCommandStrings", "0.12.0")
 	def RebootCommand = "reboot"
-	def RebootSummary = RebootCommand + " [full]"
-	def RebootBrief = (RebootSummary, "Reboots sbt and then executes the remaining commands.")
-	def RebootDetailed =
-RebootSummary + """
 
-	This command is equivalent to exiting sbt, restarting, and running the
-	  remaining commands with the exception that the JVM is not shut down.
-
-	If 'full' is specified, the boot directory (`~/.sbt/boot` by default)
-	  is deleted before restarting.  This forces an update of sbt and Scala
-	  and is useful when working with development versions of sbt or Scala."""
-
+	@deprecated("Moved to BasicCommandStrings", "0.12.0")
 	def Multi = ";"
-	def MultiBrief = (Multi + " <command> (" + Multi + " <command>)*", "Runs the provided semicolon-separated commands.")
-	def MultiDetailed =
-Multi + " command1 " + Multi + """ command2 ...
 
-	Runs the specified commands."""
-
+	@deprecated("Moved to BasicCommandStrings", "0.12.0")
 	def AppendCommand = "append"
-	def AppendLastBrief = (AppendCommand + " <command>", AppendLastDetailed)
-	def AppendLastDetailed = "Appends 'command' to list of commands to run."
-
-	val AliasCommand = "alias"
-	def AliasBrief = (AliasCommand, "Adds, removes, or prints command aliases.")
-	def AliasDetailed =
-AliasCommand + """
-
-	Prints a list of defined aliases.
-
-""" +
-AliasCommand + """ name
-
-	Prints the alias defined for `name`.
-
-""" +
-AliasCommand + """ name=value
-
-	Sets the alias `name` to `value`, replacing any existing alias with that name.
-	Whenever `name` is entered, the corresponding `value` is run.
-	If any argument is provided to `name`, it is appended as argument to `value`.
-
-""" +
-AliasCommand + """ name=
-
-	Removes the alias for `name`."""
-
-	def Discover = "discover"
-	def DiscoverBrief = (DiscoverSyntax, "Finds annotated classes and subclasses.")
-	def DiscoverSyntax = Discover + " [-module true|false] [-sub <names>] [-annot <names>]"
-	def DiscoverDetailed =
-DiscoverSyntax + """
-
-	Looks for public, concrete classes that match the requested query using the current sbt.inc.Analysis instance.
-
-	-module
-		Specifies whether modules (true) or classes (false) are found.
-		The default is classes/traits (false).
-
-	-sub
-		Specifies comma-separated class names.
-		Classes that have one or more of these classes as an ancestor are included in the resulting list.
-
-	-annot
-		Specifies comma-separated annotation names.
-		Classes with one or more of these annotations on the class or one of its non-private methods are included in the resulting list.
-"""
-
-	def CompileName = "direct-compile"
-	def CompileBrief = (CompileSyntax, "Incrementally compiles the provided sources.")
-	def CompileSyntax = CompileName + " -src <paths> [-cp <paths>] [-d <path>]"
-	def CompileDetailed =
-CompileSyntax + """
-
-	Incrementally compiles Scala and Java sources.
-
-	<paths> are explicit paths separated by the platform path separator.
-
-	The specified output path will contain the following directory structure:
-
-		scala_<version>/
-			classes/
-			cache/
-
-	Compiled classes will be written to the 'classes' directory.
-	Cached information about the compilation will be written to 'cache'.
-"""
-
-	val FailureWall = "---"
 
 	def Load = "load"
 	def LoadLabel = "a project"
@@ -308,31 +183,20 @@ CompileSyntax + """
 	def LoadProjectBrief = (LoadProject, LoadProjectDetailed)
 	def LoadProjectDetailed = "Loads the project in the current directory"
 
-	def Shell = "shell"
-	def ShellBrief = ShellDetailed
-	def ShellDetailed = "Provides an interactive prompt from which commands can be run."
+	@deprecated("Moved to State", "0.12.0")
+	val FailureWall = State.FailureWall
 
+	@deprecated("Moved to BasicCommandStrings", "0.12.0")
+	def Shell = BasicCommandStrings.Shell
+
+	@deprecated("Moved to BasicCommandStrings", "0.12.0")
 	def ClearOnFailure = "--"
+
+	@deprecated("Moved to BasicCommandStrings", "0.12.0")
 	def OnFailure = "-"
-	def OnFailureBrief = (OnFailure + " command", "Registers 'command' to run if a command fails.")
-	def OnFailureDetailed =
-OnFailure + """ command
 
-	Registers 'command' to run when a command fails to complete normally.
-
-	Only one failure command may be registered at a time, so this command
-	  replaces the previous command if there is one.
-
-	The failure command resets when it runs once, so it must be added
-	  again if desired."""
-
+	@deprecated("Moved to BasicCommandStrings", "0.12.0")
 	def IfLast = "iflast"
-	def IfLastBrief = (IfLast + " <command>", IfLastCommon)
-	def IfLastCommon = "If there are no more commands after this one, 'command' is run."
-	def IfLastDetailed =
-IfLast + """ command
-
-	""" + IfLastCommon
 
 	def InitCommand = "initialize"
 	def InitBrief = (InitCommand, "Initializes command processing.")
@@ -352,4 +216,12 @@ load-commands -base ~/.sbt/commands
 < .sbtrc
 	Runs commands from ~/.sbtrc and ./.sbtrc if they exist
 """
+
+		import java.io.File
+		import Path._
+
+	def sbtRCs(s: State): Seq[File] =
+		(Path.userHome / sbtrc) ::
+		(s.baseDir / sbtrc asFile) ::
+		Nil
 }
