+++ sbt/package.scala
@@ -4,6 +4,9 @@
 package object sbt extends sbt.std.TaskExtra with sbt.Types with sbt.ProcessExtra with sbt.impl.DependencyBuilders
 	with sbt.PathExtra with sbt.ProjectExtra with sbt.DependencyFilterExtra with sbt.BuildExtra
 {
+	@deprecated("Renamed to CommandStrings.", "0.12.0")
+	val CommandSupport = CommandStrings
+
 	@deprecated("Use SettingKey, which is a drop-in replacement.", "0.11.1")
 	type ScopedSetting[T] = SettingKey[T]
 	@deprecated("Use TaskKey, which is a drop-in replacement.", "0.11.1")
